Revert "Sprint 6 solution"
