+++ src/Epic.java
@@ -1,5 +1,3 @@
-package tasks;
-
 import java.util.ArrayList;
 import java.util.Objects;
 
@@ -20,7 +18,7 @@ public class Epic extends Task {
 
     @Override
     public String toString() {
-        return "Tasks.Epic{" +
+        return "Epic{" +
                 "name='" + getName() + '\'' +
                 ", description='" + getDescription() + '\'' +
                 ", status=" + status +
