+++ src/EpicTest.java
@@ -1,7 +1,3 @@
-import manager.InMemoryTaskManager;
-import manager.TaskManager;
-import org.junit.jupiter.api.Assertions;
-import tasks.Epic;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -11,7 +7,7 @@ class EpicTest {
     public void compareEpics() {
         Epic epic = new Epic("fs", "34hj", 2);
         Epic epic1 = new Epic("joj", "0i[", 2);
-        Assertions.assertEquals(epic, epic1);
+        assertTrue(epic.equals(epic1));
     }
 
     @Test
