+++ src/HistoryManager.java
@@ -1,13 +1,7 @@
-package manager;
-
-import tasks.Task;
-
 import java.util.List;
 
 public interface HistoryManager {
     void add(Task task);
 
     List<Task> getHistory();
-
-    void remove(int id);
 }
