+++ src/Managers.java
@@ -1,5 +1,3 @@
-package manager;
-
 public class Managers {
     public static TaskManager getDefault() {
         return new InMemoryTaskManager();
