+++ src/ManagersTest.java
@@ -1,6 +1,3 @@
-import manager.HistoryManager;
-import manager.Managers;
-import manager.TaskManager;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
