+++ src/Status.java
@@ -1,5 +1,3 @@
-package tasks;
-
 public enum Status {
     NEW,
     IN_PROGRESS,
