+++ src/Subtask.java
@@ -1,5 +1,3 @@
-package tasks;
-
 import java.util.Objects;
 
 public class Subtask extends Task {
@@ -16,7 +14,7 @@ public class Subtask extends Task {
 
     @Override
     public String toString() {
-        return "Tasks.Subtask{" +
+        return "Subtask{" +
                 "idOfEpic=" + getIdOfEpic() +
                 ", name='" + getName() + '\'' +
                 ", description='" + getDescription() + '\'' +
