+++ src/SubtaskTest.java
@@ -1,8 +1,3 @@
-import manager.InMemoryTaskManager;
-import manager.TaskManager;
-import tasks.Epic;
-import tasks.Status;
-import tasks.Subtask;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -12,7 +7,7 @@ class SubtaskTest {
     public void compareSubtasks() {
         Subtask subtask = new Subtask("gew", "ojog", 5, Status.NEW, 1);
         Subtask subtask1 = new Subtask("gew", "ojog", 5, Status.NEW, 1);
-        assertEquals(subtask, subtask1);
+        assertTrue(subtask.equals(subtask1));
     }
 
     @Test
