+++ src/TaskManager.java
@@ -1,9 +1,3 @@
-package manager;
-
-import tasks.Epic;
-import tasks.Subtask;
-import tasks.Task;
-
 import java.util.List;
 
 public interface TaskManager {
