+++ src/TaskTest.java
@@ -1,7 +1,3 @@
-import manager.InMemoryTaskManager;
-import manager.TaskManager;
-import tasks.Status;
-import tasks.Task;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
